pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["it's me"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
sqlalchemy = "^2.0.31"
asyncio = "^3.4.3"
asyncpg = "^0.29.0"
uvicorn = "^0.30.5"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
alembic = "^1.13.2"
uuid = "^1.30"
aiobotocore = "^2.13.1"
httpx = "^0.27.0"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
anyio = "^4.4.0"
pytest-tornasync = "^0.6.0.post2"
pytest-trio = "^0.8.0"
pytest-twisted = "^1.14.2"
alt-pytest-asyncio = "^0.8.1"
pytest-tornado = "^0.8.1"
werkzeug = "^3.0.3"
twisted = "^24.3.0"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


